+++ package.json
@@ -20,6 +20,7 @@
 	},
 	"type": "module",
 	"dependencies": {
+		"cookie": "^0.5.0",
 		"dotenv": "^16.0.1",
 		"svelte-local-storage-store": "^0.2.6"
 	}
