+++ src/routes/+page.js
@@ -2,7 +2,7 @@ import { redirect } from "@sveltejs/kit";
 
 export async function load({ parent }) {
 	const { desktop } = await parent();
-	if (!desktop) {
+	if (desktop === false) {
 		throw redirect(302, "/mobile");
 	}
 }
