+++ src/routes/+page.svelte
@@ -3,8 +3,6 @@
 	import Stats from "$lib/components/Stats.svelte";
 	import { page } from "$app/stores";
 
-	console.log($page.url.hash);
-
 	let displayStats = false;
 </script>
